refactor: extract clampValues to a common utilities module
