+++ src/utilities.ts
@@ -1,0 +1,11 @@
+export function clampValues(value: number, minValue: number, maxValue: number): number {
+  if (value < minValue) {
+    return minValue;
+  }
+
+  if (value > maxValue) {
+    return maxValue;
+  }
+
+  return value;
+}
